pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "resume-tailor-crew"
version = "0.1.0"
description = "Enterprise-grade CrewAI system for intelligent resume tailoring based on job descriptions"
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = { text = "MIT" }
authors = [
    { name = "Resume Tailor Team", email = "team@resumetailor.dev" }
]
keywords = [
    "crewai",
    "resume",
    "tailoring",
    "job-matching",
    "ai-agents",
    "nlp"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Text Processing",
]

dependencies = [
    "crewai[google-genai]>=0.1.0",
    "crewai-tools>=0.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "structlog>=25.0.0",
    "tenacity>=8.2.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "ruff>=0.0.280",
    "mypy>=1.5.0",
    "pylint>=2.17.0",
    "isort>=5.12.0",
    "pre-commit>=3.3.0",
]

llm = [
    "openai>=1.0.0",
    "anthropic>=0.7.0",
]

document-processing = [
    "docling>=1.0.0",        # Tier 1: Advanced PDF parsing (multi-column, tables)
    "mineru>=0.1.0",         # Tier 2: Fallback PDF/DOCX processor
    "PyPDF2>=3.0.0",         # Tier 3: Basic PDF text extraction
    "python-docx>=1.0.0",    # Tier 3: Basic DOCX text extraction
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.2.0",
    "mkdocstrings>=0.23.0",
]

[project.urls]
Repository = "https://github.com/your-org/resume-tailor-crew"
Documentation = "https://resume-tailor-crew.readthedocs.io"
"Issue Tracker" = "https://github.com/your-org/resume-tailor-crew/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/agents", "src/config", "src/core", "src/data_models", "src/tools"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "C",      # flake8-comprehensions
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by ruff format)
    "C901",   # function is too complex
    "B905",   # zip() without an explicit strict= parameter (Python 3.10+)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
known_first_party = ["resume_tailor_crew"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true

[[tool.mypy.overrides]]
module = "crewai.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--strict-markers --tb=short -v"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

[tool.coverage.run]
source = ["src/resume_tailor_crew"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
precision = 2

